+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Changed
   - (**MAJOR**) All custom webviews are now using the [vscode-webview-ui-toolkit](https://github.com/microsoft/vscode-webview-ui-toolkit)
+- Added
+  - Copying bucket and object keys into clipboard
 
 ## [2.4.1] - 2023-03-10